feat(llvm): support `lw/sw` and `lb/sb` for i8/i32 load/stores (#953)
This PR adds `lw/sw`(i32)  as well as `lb/sb` (i8)  lowerings.

On proofs: These aren't fully proven yet but it is analogous to the rest
of memory (i64). Proves and follow-up exec test as a lightweight check
will be added in the future + someone else working on the memory proofs.

Where these lowering come from: LLVM's RISC-V backend selects integer
loads/stores by byte width — i8/i32/i64 → LB/LW/LD and SB/SW/SD (it only
uses the zero-extending LBU/LWU when the loaded value is explicitly
zero-extended). We mirror that with the plain signed loads and I think
we should implement the LBU/LWU lowering as a rewrite in the future?
Because it needs to also account for having a zero-extension after the
load which can be pattern-matched.

---------

+++ Test/Passes/InstructionSelection/RISCV64/load_invalid.mlir
@@ -3,8 +3,9 @@
 "builtin.module"() ({
     "func.func"()  <{function_type = (!llvm.ptr) -> ()}> ({
     ^bb0(%a: !llvm.ptr):
-        %val = "llvm.load"(%a) : (!llvm.ptr) -> i32
-        // CHECK: {{.*}} = "llvm.load"({{.*}}) <{"access_groups" = [], "alias_scopes" = [], "alignment" = 0 : i64, "noalias_scopes" = [], "tbaa" = []}> : (!llvm.ptr) -> i32
+        // i16 loads are not supported (only i64/i32/i8), so this stays un-lowered.
+        %val = "llvm.load"(%a) : (!llvm.ptr) -> i16
+        // CHECK: {{.*}} = "llvm.load"({{.*}}) <{"access_groups" = [], "alias_scopes" = [], "alignment" = 0 : i64, "noalias_scopes" = [], "tbaa" = []}> : (!llvm.ptr) -> i16
         "func.return"() : () -> ()
     }) : () -> ()
 }) : () -> ()
